Fix build tags
